add op_mapper test directory
